requirements.txt
azure-core==1.9.0
azure-eventhub==5.2.0
python-dotenv==0.15.0